Marge and rename test.rs to test_old.rs bc i'm not fixing that rn
